modules/location_producer_microservice/requirements.txt
grpcio==1.47.2
grpcio-tools==1.47.2
kafka-python==2.0.2